+++ app/ide-desktop/lib/dashboard/src/authentication/src/authentication/providers/auth.tsx
@@ -3,31 +3,31 @@
  * Provides an `AuthProvider` component that wraps the entire application, and a `useAuth` hook that
  * can be used from any React component to access the currently logged-in user's session data. The
  * hook also provides methods for registering a user, logging in, logging out, etc. */
-import * as react from "react";
-import * as router from "react-router-dom";
-import toast from "react-hot-toast";
+import * as react from 'react'
+import * as router from 'react-router-dom'
+import toast from 'react-hot-toast'
 
-import * as app from "../../components/app";
-import * as authServiceModule from "../service";
-import * as backendService from "../../dashboard/service";
-import * as errorModule from "../../error";
-import * as loggerProvider from "../../providers/logger";
-import * as sessionProvider from "./session";
+import * as app from '../../components/app'
+import * as authServiceModule from '../service'
+import * as backendService from '../../dashboard/service'
+import * as errorModule from '../../error'
+import * as loggerProvider from '../../providers/logger'
+import * as sessionProvider from './session'
 
 // =================
 // === Constants ===
 // =================
 
 const MESSAGES = {
-  signUpSuccess: "We have sent you an email with further instructions!",
-  confirmSignUpSuccess: "Your account has been confirmed! Please log in.",
-  setUsernameSuccess: "Your username has been set!",
-  signInWithPasswordSuccess: "Successfully logged in!",
-  forgotPasswordSuccess: "We have sent you an email with further instructions!",
-  resetPasswordSuccess: "Successfully reset password!",
-  signOutSuccess: "Successfully logged out!",
-  pleaseWait: "Please wait...",
-} as const;
+    signUpSuccess: 'We have sent you an email with further instructions!',
+    confirmSignUpSuccess: 'Your account has been confirmed! Please log in.',
+    setUsernameSuccess: 'Your username has been set!',
+    signInWithPasswordSuccess: 'Successfully logged in!',
+    forgotPasswordSuccess: 'We have sent you an email with further instructions!',
+    resetPasswordSuccess: 'Successfully reset password!',
+    signOutSuccess: 'Successfully logged out!',
+    pleaseWait: 'Please wait...',
+} as const
 
 // =============
 // === Types ===
@@ -35,19 +35,19 @@ const MESSAGES = {
 
 // === UserSession ===
 
-export type UserSession = FullUserSession | PartialUserSession;
+export type UserSession = FullUserSession | PartialUserSession
 
 /** Object containing the currently signed-in user's session data. */
 export interface FullUserSession {
-  /** A discriminator for TypeScript to be able to disambiguate between this interface and other
-   * `UserSession` variants. */
-  variant: "full";
-  /** User's JSON Web Token (JWT), used for authenticating and authorizing requests to the API. */
-  accessToken: string;
-  /** User's email address. */
-  email: string;
-  /** User's organization information. */
-  organization: backendService.Organization;
+    /** A discriminator for TypeScript to be able to disambiguate between this interface and other
+     * `UserSession` variants. */
+    variant: 'full'
+    /** User's JSON Web Token (JWT), used for authenticating and authorizing requests to the API. */
+    accessToken: string
+    /** User's email address. */
+    email: string
+    /** User's organization information. */
+    organization: backendService.Organization
 }
 
 /** Object containing the currently signed-in user's session data, if the user has not yet set their
@@ -57,13 +57,13 @@ export interface FullUserSession {
  * their account. Otherwise, this type is identical to the `Session` type. This type should ONLY be
  * used by the `SetUsername` component. */
 export interface PartialUserSession {
-  /** A discriminator for TypeScript to be able to disambiguate between this interface and other
-   * `UserSession` variants. */
-  variant: "partial";
-  /** User's JSON Web Token (JWT), used for authenticating and authorizing requests to the API. */
-  accessToken: string;
-  /** User's email address. */
-  email: string;
+    /** A discriminator for TypeScript to be able to disambiguate between this interface and other
+     * `UserSession` variants. */
+    variant: 'partial'
+    /** User's JSON Web Token (JWT), used for authenticating and authorizing requests to the API. */
+    accessToken: string
+    /** User's email address. */
+    email: string
 }
 
 // ===================
@@ -78,27 +78,19 @@ export interface PartialUserSession {
  *
  * See {@link Cognito} for details on each of the authentication functions. */
 interface AuthContextType {
-  signUp: (email: string, password: string) => Promise<void>;
-  confirmSignUp: (email: string, code: string) => Promise<void>;
-  setUsername: (
-    accessToken: string,
-    username: string,
-    email: string
-  ) => Promise<void>;
-  signInWithGoogle: () => Promise<void>;
-  signInWithGitHub: () => Promise<void>;
-  signInWithPassword: (email: string, password: string) => Promise<void>;
-  forgotPassword: (email: string) => Promise<void>;
-  resetPassword: (
-    email: string,
-    code: string,
-    password: string
-  ) => Promise<void>;
-  signOut: () => Promise<void>;
-  /** Session containing the currently authenticated user's authentication information.
-   *
-   * If the user has not signed in, the session will be `null`. */
-  session: UserSession | null;
+    signUp: (email: string, password: string) => Promise<void>
+    confirmSignUp: (email: string, code: string) => Promise<void>
+    setUsername: (accessToken: string, username: string, email: string) => Promise<void>
+    signInWithGoogle: () => Promise<void>
+    signInWithGitHub: () => Promise<void>
+    signInWithPassword: (email: string, password: string) => Promise<void>
+    forgotPassword: (email: string) => Promise<void>
+    resetPassword: (email: string, code: string, password: string) => Promise<void>
+    signOut: () => Promise<void>
+    /** Session containing the currently authenticated user's authentication information.
+     *
+     * If the user has not signed in, the session will be `null`. */
+    session: UserSession | null
 }
 
 // Eslint doesn't like headings.
@@ -126,192 +118,186 @@ interface AuthContextType {
  * So changing the cast would provide no safety guarantees, and would require us to introduce null
  * checks everywhere we use the context. */
 // eslint-disable-next-line no-restricted-syntax
-const AuthContext = react.createContext<AuthContextType>({} as AuthContextType);
+const AuthContext = react.createContext<AuthContextType>({} as AuthContextType)
 
 // ====================
 // === AuthProvider ===
 // ====================
 
 export interface AuthProviderProps {
-  authService: authServiceModule.AuthService;
-  /** Callback to execute once the user has authenticated successfully. */
-  onAuthenticated: () => void;
-  children: react.ReactNode;
+    authService: authServiceModule.AuthService
+    /** Callback to execute once the user has authenticated successfully. */
+    onAuthenticated: () => void
+    children: react.ReactNode
 }
 
 export function AuthProvider(props: AuthProviderProps) {
-  const { authService, children } = props;
-  const { cognito } = authService;
-  const { session } = sessionProvider.useSession();
-  const logger = loggerProvider.useLogger();
-  const navigate = router.useNavigate();
-  const onAuthenticated = react.useCallback(props.onAuthenticated, []);
-  const [initialized, setInitialized] = react.useState(false);
-  const [userSession, setUserSession] = react.useState<UserSession | null>(
-    null
-  );
+    const { authService, children } = props
+    const { cognito } = authService
+    const { session } = sessionProvider.useSession()
+    const logger = loggerProvider.useLogger()
+    const navigate = router.useNavigate()
+    const onAuthenticated = react.useCallback(props.onAuthenticated, [])
+    const [initialized, setInitialized] = react.useState(false)
+    const [userSession, setUserSession] = react.useState<UserSession | null>(null)
 
-  /** Fetch the JWT access token from the session via the AWS Amplify library.
-   *
-   * When invoked, retrieves the access token (if available) from the storage method chosen when
-   * Amplify was configured (e.g. local storage). If the token is not available, return `undefined`.
-   * If the token has expired, automatically refreshes the token and returns the new token. */
-  react.useEffect(() => {
-    const fetchSession = async () => {
-      if (session.none) {
-        setInitialized(true);
-        setUserSession(null);
-      } else {
-        const { accessToken, email } = session.val;
+    /** Fetch the JWT access token from the session via the AWS Amplify library.
+     *
+     * When invoked, retrieves the access token (if available) from the storage method chosen when
+     * Amplify was configured (e.g. local storage). If the token is not available, return `undefined`.
+     * If the token has expired, automatically refreshes the token and returns the new token. */
+    react.useEffect(() => {
+        const fetchSession = async () => {
+            if (session.none) {
+                setInitialized(true)
+                setUserSession(null)
+            } else {
+                const { accessToken, email } = session.val
 
-        const backend = backendService.createBackend(accessToken, logger);
-        const organization = await backend.getUser();
-        let newUserSession: UserSession;
-        if (!organization) {
-          newUserSession = {
-            variant: "partial",
-            email,
-            accessToken,
-          };
-        } else {
-          newUserSession = {
-            variant: "full",
-            email,
-            accessToken,
-            organization,
-          };
+                const backend = backendService.createBackend(accessToken, logger)
+                const organization = await backend.getUser()
+                let newUserSession: UserSession
+                if (!organization) {
+                    newUserSession = {
+                        variant: 'partial',
+                        email,
+                        accessToken,
+                    }
+                } else {
+                    newUserSession = {
+                        variant: 'full',
+                        email,
+                        accessToken,
+                        organization,
+                    }
 
-          /** Execute the callback that should inform the Electron app that the user has logged in.
-           * This is done to transition the app from the authentication/dashboard view to the IDE. */
-          onAuthenticated();
+                    /** Execute the callback that should inform the Electron app that the user has logged in.
+                     * This is done to transition the app from the authentication/dashboard view to the IDE. */
+                    onAuthenticated()
+                }
+
+                setUserSession(newUserSession)
+                setInitialized(true)
+            }
         }
 
-        setUserSession(newUserSession);
-        setInitialized(true);
-      }
-    };
+        fetchSession().catch(error => {
+            if (isUserFacingError(error)) {
+                toast.error(error.message)
+            } else {
+                logger.error(error)
+            }
+        })
+    }, [session])
 
-    fetchSession().catch((error) => {
-      if (isUserFacingError(error)) {
-        toast.error(error.message);
-      } else {
-        logger.error(error);
-      }
-    });
-  }, [session]);
-
-  const withLoadingToast =
-    <T extends unknown[]>(action: (...args: T) => Promise<void>) =>
-    async (...args: T) => {
-      const loadingToast = toast.loading(MESSAGES.pleaseWait);
-      try {
-        await action(...args);
-      } finally {
-        toast.dismiss(loadingToast);
-      }
-    };
-
-  const signUp = (username: string, password: string) =>
-    cognito.signUp(username, password).then((result) => {
-      if (result.ok) {
-        toast.success(MESSAGES.signUpSuccess);
-      } else {
-        toast.error(result.val.message);
-      }
-    });
-
-  const confirmSignUp = async (email: string, code: string) =>
-    cognito.confirmSignUp(email, code).then((result) => {
-      if (result.err) {
-        switch (result.val.kind) {
-          case "UserAlreadyConfirmed":
-            break;
-          default:
-            throw new errorModule.UnreachableCaseError(result.val.kind);
-        }
-      }
-
-      toast.success(MESSAGES.confirmSignUpSuccess);
-      navigate(app.LOGIN_PATH);
-    });
-
-  const signInWithPassword = async (email: string, password: string) =>
-    cognito.signInWithPassword(email, password).then((result) => {
-      if (result.ok) {
-        toast.success(MESSAGES.signInWithPasswordSuccess);
-      } else {
-        if (result.val.kind === "UserNotFound") {
-          navigate(app.REGISTRATION_PATH);
+    const withLoadingToast =
+        <T extends unknown[]>(action: (...args: T) => Promise<void>) =>
+        async (...args: T) => {
+            const loadingToast = toast.loading(MESSAGES.pleaseWait)
+            try {
+                await action(...args)
+            } finally {
+                toast.dismiss(loadingToast)
+            }
         }
 
-        toast.error(result.val.message);
-      }
-    });
+    const signUp = (username: string, password: string) =>
+        cognito.signUp(username, password).then(result => {
+            if (result.ok) {
+                toast.success(MESSAGES.signUpSuccess)
+            } else {
+                toast.error(result.val.message)
+            }
+        })
 
-  const setUsername = async (
-    accessToken: string,
-    username: string,
-    email: string
-  ) => {
-    const body: backendService.SetUsernameRequestBody = {
-      userName: username,
-      userEmail: email,
-    };
+    const confirmSignUp = async (email: string, code: string) =>
+        cognito.confirmSignUp(email, code).then(result => {
+            if (result.err) {
+                switch (result.val.kind) {
+                    case 'UserAlreadyConfirmed':
+                        break
+                    default:
+                        throw new errorModule.UnreachableCaseError(result.val.kind)
+                }
+            }
 
-    /** TODO [NP]: https://github.com/enso-org/cloud-v2/issues/343
-     * The API client is reinitialised on every request. That is an inefficient way of usage.
-     * Fix it by using React context and implementing it as a singleton. */
-    const backend = backendService.createBackend(accessToken, logger);
+            toast.success(MESSAGES.confirmSignUpSuccess)
+            navigate(app.LOGIN_PATH)
+        })
 
-    await backend.setUsername(body);
-    navigate(app.DASHBOARD_PATH);
-    toast.success(MESSAGES.setUsernameSuccess);
-  };
+    const signInWithPassword = async (email: string, password: string) =>
+        cognito.signInWithPassword(email, password).then(result => {
+            if (result.ok) {
+                toast.success(MESSAGES.signInWithPasswordSuccess)
+            } else {
+                if (result.val.kind === 'UserNotFound') {
+                    navigate(app.REGISTRATION_PATH)
+                }
 
-  const forgotPassword = async (email: string) =>
-    cognito.forgotPassword(email).then((result) => {
-      if (result.ok) {
-        toast.success(MESSAGES.forgotPasswordSuccess);
-        navigate(app.RESET_PASSWORD_PATH);
-      } else {
-        toast.error(result.val.message);
-      }
-    });
+                toast.error(result.val.message)
+            }
+        })
 
-  const resetPassword = async (email: string, code: string, password: string) =>
-    cognito.forgotPasswordSubmit(email, code, password).then((result) => {
-      if (result.ok) {
-        toast.success(MESSAGES.resetPasswordSuccess);
-        navigate(app.LOGIN_PATH);
-      } else {
-        toast.error(result.val.message);
-      }
-    });
+    const setUsername = async (accessToken: string, username: string, email: string) => {
+        const body: backendService.SetUsernameRequestBody = {
+            userName: username,
+            userEmail: email,
+        }
 
-  const signOut = () =>
-    cognito.signOut().then(() => {
-      toast.success(MESSAGES.signOutSuccess);
-    });
+        /** TODO [NP]: https://github.com/enso-org/cloud-v2/issues/343
+         * The API client is reinitialised on every request. That is an inefficient way of usage.
+         * Fix it by using React context and implementing it as a singleton. */
+        const backend = backendService.createBackend(accessToken, logger)
 
-  const value = {
-    signUp: withLoadingToast(signUp),
-    confirmSignUp: withLoadingToast(confirmSignUp),
-    setUsername,
-    signInWithGoogle: cognito.signInWithGoogle.bind(cognito),
-    signInWithGitHub: cognito.signInWithGitHub.bind(cognito),
-    signInWithPassword: withLoadingToast(signInWithPassword),
-    forgotPassword: withLoadingToast(forgotPassword),
-    resetPassword: withLoadingToast(resetPassword),
-    signOut,
-    session: userSession,
-  };
+        await backend.setUsername(body)
+        navigate(app.DASHBOARD_PATH)
+        toast.success(MESSAGES.setUsernameSuccess)
+    }
 
-  return (
-    <AuthContext.Provider value={value}>
-      {/* Only render the underlying app after we assert for the presence of a current user. */}
-      {initialized && children}
-    </AuthContext.Provider>
-  );
+    const forgotPassword = async (email: string) =>
+        cognito.forgotPassword(email).then(result => {
+            if (result.ok) {
+                toast.success(MESSAGES.forgotPasswordSuccess)
+                navigate(app.RESET_PASSWORD_PATH)
+            } else {
+                toast.error(result.val.message)
+            }
+        })
+
+    const resetPassword = async (email: string, code: string, password: string) =>
+        cognito.forgotPasswordSubmit(email, code, password).then(result => {
+            if (result.ok) {
+                toast.success(MESSAGES.resetPasswordSuccess)
+                navigate(app.LOGIN_PATH)
+            } else {
+                toast.error(result.val.message)
+            }
+        })
+
+    const signOut = () =>
+        cognito.signOut().then(() => {
+            toast.success(MESSAGES.signOutSuccess)
+        })
+
+    const value = {
+        signUp: withLoadingToast(signUp),
+        confirmSignUp: withLoadingToast(confirmSignUp),
+        setUsername,
+        signInWithGoogle: cognito.signInWithGoogle.bind(cognito),
+        signInWithGitHub: cognito.signInWithGitHub.bind(cognito),
+        signInWithPassword: withLoadingToast(signInWithPassword),
+        forgotPassword: withLoadingToast(forgotPassword),
+        resetPassword: withLoadingToast(resetPassword),
+        signOut,
+        session: userSession,
+    }
+
+    return (
+        <AuthContext.Provider value={value}>
+            {/* Only render the underlying app after we assert for the presence of a current user. */}
+            {initialized && children}
+        </AuthContext.Provider>
+    )
 }
 
 /** Type of an error containing a `string`-typed `message` field.
@@ -319,13 +305,13 @@ export function AuthProvider(props: AuthProviderProps) {
  * Many types of errors fall into this category. We use this type to check if an error can be safely
  * displayed to the user. */
 interface UserFacingError {
-  /** The user-facing error message. */
-  message: string;
+    /** The user-facing error message. */
+    message: string
 }
 
 /** Returns `true` if the value is a {@link UserFacingError}. */
 function isUserFacingError(value: unknown): value is UserFacingError {
-  return typeof value === "object" && value != null && "message" in value;
+    return typeof value === 'object' && value != null && 'message' in value
 }
 
 // ===============
@@ -337,7 +323,7 @@ function isUserFacingError(value: unknown): value is UserFacingError {
  * Only the hook is exported, and not the context, because we only want to use the hook directly and
  * never the context component. */
 export function useAuth() {
-  return react.useContext(AuthContext);
+    return react.useContext(AuthContext)
 }
 
 // =======================
@@ -346,13 +332,13 @@ export function useAuth() {
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 export function ProtectedLayout() {
-  const { session } = useAuth();
+    const { session } = useAuth()
 
-  if (!session) {
-    return <router.Navigate to={app.LOGIN_PATH} />;
-  } else {
-    return <router.Outlet context={session} />;
-  }
+    if (!session) {
+        return <router.Navigate to={app.LOGIN_PATH} />
+    } else {
+        return <router.Outlet context={session} />
+    }
 }
 
 // ===================
@@ -361,15 +347,15 @@ export function ProtectedLayout() {
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 export function GuestLayout() {
-  const { session } = useAuth();
+    const { session } = useAuth()
 
-  if (session?.variant === "partial") {
-    return <router.Navigate to={app.SET_USERNAME_PATH} />;
-  } else if (session?.variant === "full") {
-    return <router.Navigate to={app.DASHBOARD_PATH} />;
-  } else {
-    return <router.Outlet />;
-  }
+    if (session?.variant === 'partial') {
+        return <router.Navigate to={app.SET_USERNAME_PATH} />
+    } else if (session?.variant === 'full') {
+        return <router.Navigate to={app.DASHBOARD_PATH} />
+    } else {
+        return <router.Outlet />
+    }
 }
 
 // =============================
@@ -377,7 +363,7 @@ export function GuestLayout() {
 // =============================
 
 export function usePartialUserSession() {
-  return router.useOutletContext<PartialUserSession>();
+    return router.useOutletContext<PartialUserSession>()
 }
 
 // ==========================
@@ -385,5 +371,5 @@ export function usePartialUserSession() {
 // ==========================
 
 export function useFullUserSession() {
-  return router.useOutletContext<FullUserSession>();
+    return router.useOutletContext<FullUserSession>()
 }