+++ std-bits/table/src/main/java/org/enso/table/data/column/builder/object/StringBuilder.java
@@ -2,6 +2,8 @@ package org.enso.table.data.column.builder.object;
 
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.column.storage.StringStorage;
+import org.enso.table.data.column.storage.type.StorageType;
+import org.enso.table.data.column.storage.type.TextType;
 
 /** A builder for string columns. */
 public class StringBuilder extends TypedBuilderImpl<String> {
@@ -15,8 +17,8 @@ public class StringBuilder extends TypedBuilderImpl<String> {
   }
 
   @Override
-  public int getType() {
-    return Storage.Type.STRING;
+  public StorageType getType() {
+    return TextType.VARIABLE_LENGTH;
   }
 
   @Override