+++ std-bits/table/src/main/java/org/enso/table/data/column/builder/object/TimeOfDayBuilder.java
@@ -2,6 +2,8 @@ package org.enso.table.data.column.builder.object;
 
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.column.storage.TimeOfDayStorage;
+import org.enso.table.data.column.storage.type.StorageType;
+import org.enso.table.data.column.storage.type.TimeOfDayType;
 
 import java.time.LocalTime;
 
@@ -17,8 +19,8 @@ public class TimeOfDayBuilder extends TypedBuilderImpl<LocalTime> {
   }
 
   @Override
-  public int getType() {
-    return Storage.Type.TIME_OF_DAY;
+  public StorageType getType() {
+    return TimeOfDayType.INSTANCE;
   }
 
   @Override