+++ CHANGELOG.md
@@ -370,6 +370,7 @@
   `use_regex` flag.][5959]
 - [Removed many regex compile flags from `split`; added `only_first` and
   `use_regex` flag.][6116]
+- [Implemented proper support for Value Types in the Table library.][6073]
 
 [debug-shortcuts]:
   https://github.com/enso-org/enso/blob/develop/app/gui/docs/product/shortcuts.md#debug
@@ -559,6 +560,7 @@
 [5705]: https://github.com/enso-org/enso/pull/5705
 [5959]: https://github.com/enso-org/enso/pull/5959
 [6116]: https://github.com/enso-org/enso/pull/6116
+[6073]: https://github.com/enso-org/enso/pull/6073
 
 #### Enso Compiler